% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{brand_colors}
\alias{brand_colors}
\title{Function to extract colors as hex codes by name}
\usage{
brand_colors(...)
}
\arguments{
\item{...}{Character names of color}
}
\description{
Function to extract colors as hex codes by name
}
