% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{brand_info}
\alias{brand_info}
\title{Prints out available named colors and palettes}
\usage{
brand_info(...)
}
\arguments{
\item{...}{Character names of color}
}
\description{
Prints out available named colors and palettes
}
