% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{brand_palettes}
\alias{brand_palettes}
\title{Function to extract lists of palette colors by name}
\usage{
brand_palettes(...)
}
\arguments{
\item{...}{Character name of palette}
}
\description{
Function to extract lists of palette colors by name
}
