% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{rancho_cols}
\alias{rancho_cols}
\title{Function to extract colors as hex codes by name}
\usage{
rancho_cols(...)
}
\arguments{
\item{...}{Character names of color}
}
\description{
Function to extract colors as hex codes by name
}
