% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{rancho_pal}
\alias{rancho_pal}
\title{Return function to interpolate a color palette}
\usage{
rancho_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in rancho_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a color palette
}
